Update generic file reader to use numpy
